consolidate tests for 3 types
